utils: Move FileMatcher from core to common
The code is not specific to ORT and can be useful for other modules.
