+++ model/src/main/kotlin/utils/RootLicenseMatcher.kt
@@ -23,7 +23,7 @@ import java.io.File
 
 import org.ossreviewtoolkit.model.LicenseFinding
 import org.ossreviewtoolkit.model.config.LicenseFilenamePatterns
-import org.ossreviewtoolkit.utils.core.FileMatcher
+import org.ossreviewtoolkit.utils.common.FileMatcher
 import org.ossreviewtoolkit.utils.core.getAllAncestorDirectories
 
 /**
