+++ utils/common/src/main/kotlin/FileMatcher.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.utils.core
+package org.ossreviewtoolkit.utils.common
 
 import org.springframework.util.AntPathMatcher
