+++ utils/core/build.gradle.kts
@@ -26,7 +26,6 @@ val log4jApiKotlinVersion: String by project
 val mockkVersion: String by project
 val okhttpVersion: String by project
 val semverVersion: String by project
-val springCoreVersion: String by project
 val xzVersion: String by project
 
 plugins {
@@ -49,7 +48,6 @@ dependencies {
     implementation("org.jetbrains.kotlin:kotlin-reflect")
     implementation("org.jetbrains.kotlin:kotlin-scripting-jsr223")
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:$kotlinxCoroutinesVersion")
-    implementation("org.springframework:spring-core:$springCoreVersion")
     implementation("org.tukaani:xz:$xzVersion")
 
     testImplementation("io.mockk:mockk:$mockkVersion")
